Uniform column selection now works!
When "Uniform column toggle" is active, a selection of a column in a block, is
propagated to the correspondent column of other blocks. Same for unselection.
Need to add the column reset when uniform column is activated (now reset only
for the deactivation).
Plot is not working for some reason, I'll take care now!