Add rectangle for zoom, double click to reset zoom
Finally we can zoom parts of the plot and reset it with double click of
every button. Also tryed the first version of scroll zoom (to be adjusted)
and first version of pan with right click but also this is not stable and
a bit lagging... will work on it!
In the meantime started a dedicated project on chatgpt and trying to add
subplot settings in the right sidebar, that's the reason for this commit
hahaha for the moment the situation is stable yuk yuk!