Add basig subplot settings, applyed to selected subplot
After clicking on a subplot, on the right sidebar are shown the applyed settings
You can show or not the legend, grid, set subplot title, x label, y label.
Text is applyed on "changed", very heavy choice: will move to "focus-out-event"
but too complicated for the moement. I go to sleep, yawn