Add save widget and cleaned some file_selector chatgpt mess
In the save widget we recycle the folder selector for choosing the "saving directory"
and saving it on the config file.
Actually only *.png format supported but .svg and .jpg coming soon.
If the file name is not set, the plotted file name is used.
If a file is already present, we add "_{next_index}" to the name.